Blocking actions when camera is moving.
There are only 2 sources of camera movement, previewing and 2-images-selected.

I do not deal with 2-images-selected scenario because we have yet to settle the routine for "how to select an interpolation and send it to history panel".

What I've done is to;

Before:
On clicking some thumbnail on left panel, stage.CLEAN -> stage.PREVIEWED (immediately)
After:
stage.CLEAN -> stage.BLOCKED (immediately)
stage.BLOCKED -> stage.PREVIEWED (right when sidebar on the right pops out)